+++ gwcelery/conf/production.py
@@ -13,8 +13,8 @@ expose_to_public = True
 """Set to True if events meeting the public alert threshold really should be
 exposed to the public."""
 
-lvalert_host = 'lvalert.cgca.uwm.edu'
-"""LVAlert host."""
+igwn_alert_group = 'gracedb'
+"""IGWN alert group."""
 
 gracedb_host = 'gracedb.ligo.org'
 """GraceDB host."""