+++ gwcelery/tests/test_tasks_igwn_alert.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 import lxml
 import pytest
 
-from ..tasks import lvalert
+from ..tasks import igwn_alert
 from . import data
 
 
@@ -52,10 +52,9 @@ def test_handle_messages(mock_superevents_handle, mock_get_event,
     alert = json.loads(payload)
     alert['object']['self'] = \
         alert['object']['self'].replace('gracedb.ligo.org', 'gracedb.invalid')
-    payload = json.dumps(alert)
 
     # Run function under test
-    lvalert.handler.dispatch(node, payload)
+    igwn_alert.handler.dispatch(node, alert)
     mock_superevents_handle.assert_called_once()
 
 
@@ -71,13 +70,12 @@ def test_handle_messages_wrong_server(mock_superevents_handle,
     alert = json.loads(payload)
     alert['object']['self'] = \
         alert['object']['self'].replace('gracedb.ligo.org', 'gracedb2.invalid')
-    payload = json.dumps(alert)
 
     # Run function under test
     caplog.set_level(logging.WARNING)
-    lvalert.handler.dispatch(node, payload)
+    igwn_alert.handler.dispatch(node, alert)
     record, *_ = caplog.records
-    assert record.message == ('ignoring LVAlert message because it is '
+    assert record.message == ('ignoring IGWN alert message because it is '
                               'intended for GraceDB server '
                               'https://gracedb2.invalid/api/, but we are set '
                               'up for server https://gracedb.invalid/api/')
@@ -95,11 +93,10 @@ def test_handle_messages_no_self_link(mock_superevents_handle,
     # Manipulate alert content
     alert = json.loads(payload)
     del alert['object']['self']
-    payload = json.dumps(alert)
 
     # Run function under test
     caplog.set_level(logging.ERROR)
-    lvalert.handler.dispatch(node, payload)
+    igwn_alert.handler.dispatch(node, alert)
     record, = caplog.records
-    assert 'LVAlert message does not contain an API URL' in record.message
+    assert 'IGWN alert message does not contain an API URL' in record.message
     mock_superevents_handle.assert_not_called()